### 2020-07-04 #2437 @whedon
[PRE REVIEW]: oolong: An R package for validating automated content analysis tools
**Submitting author:** @chainsawriot (<a href="http://orcid.org/0000-0002-6232-7530">Chung-hong Chan</a>)
**Repository:** <a href="https://github.com/chainsawriot/oolong" target ="_blank">https://github.com/chainsawriot/oolong</a>
**Version:** 0.3.8
