### 2020-07-09 #2437 @kakiac
👋  @chainsawriot thanks for your submission to JOSS.

Your submission fulfils our submission criteria (https://joss.readthedocs.io/en/latest/submitting.html), so I will be contacting your suggested reviewers shortly.

In the meantime, can you please include a "Statement of need" section in your paper?

Many thanks and I will be in touch!
