### 2020-07-09 #2437 @kakiac
👋  @pdwaggoner, @mbod would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
