### 2020-07-09 91382a4
Add statement of need to the paper. ref openjournals/joss-reviews#2437
