### 2020-07-09 #2437 @akbaritabar
Dear @kakiac , thanks a lot for giving me the chance. I am really sorry. I read the paper and checked the review guidelines, but I don't feel confident enough to evaluate the technical aspects of this package. The paper is well written, but I cannot be a good judge to say anything about the functionality as it is not my area of expertise.

It is a pity, as it could have been my first JOSS review :( and I am looking forward to help.
