### 2020-07-15 #2461 @pdwaggoner
Hi @kakiac , @mbod , and @chainsawriot - Thanks for the package and for the chance to review the paper and software. For the most part this looks like a good, simple package and paper. Still, I think there are a few revisions needed prior to publication at JOSS:

- There is no statement of or clear path to community contribution (in general or for bug reports) either in the **paper** or in the **GH repo** (though it's buried in the `DESCRIPTION` file via `BugReports`). Some note to users on how to contribute to the package in these *two* locations would be nice.

- In the **paper**, there is no "state of the field". Though a few other packages are cited, though in reference to those supported by `oolong`, there is currently only a single paragraph at the outset of the paper that gives a few sentences on what validation is according to two papers (Chang et al. 2009 and Song et al. 2020). A few more paragraphs that, a) go into more depth on validation and why these techniques over the *many* other possible approaches to validation, b) detail the packages that currently exist on related topics and approaches to validation of content analytic models, and thus c) defend why this package (`oolong`) is usefully filling a real gap in the literature/software, would all be quite useful.

- Related to the previous point, in general, I found myself thinking about many other ways to validate commonly fit content analytic models. For example, the most simple and widely used validation approaches to topic models are to check coherence (and also perplexity) score(s), as a core issue with topic models is whether `k` is an appropriate characterization of the corpus. As topic models are usually unsupervised, this is almost always the first step in validating these models. Though this specific flavor of validation doesn't seem to be at the core of this package (by assumption), I wondered why not? This question could likely be satisfied by either a more thorough justification for the selected techniques mentioned in the previous point, or by expanding the package to include other prominent validation techniques for content analysis. But at a minimum, some nod to the partial selection of validation techniques (beyond noting that the authors of those two papers suggested them) would be extremely valuable, deepen the reliability of the software, as well as sharpen the contribution of this piece more broadly.

Other than these issues, the paper and software look good. I look forward to seeing the revisions and then completing my review. Thanks all!
