### 2020-07-18 chainsawriot/oolong#31 @chainsawriot
Revision related to openjournals/joss-reviews#2461
Ref openjournals/joss-reviews#2461

Comments from @pdwaggoner

- [x] software / paper: statement of or clear path to community contribution
- [x] paper: state of the field
- [x] paper: Query about why not coherence / perplexity
