+++ README.Rmd
@@ -27,4 +27,16 @@ The goal of oolong ^[/ˈuːlʊŋ/ 烏龍, literally means "Dark Dragon", is a se
 
 Please refer to the [overview](overview_gh.md) for an introduction to this package. If you use BTM, please refer to the [BTM Vignette](btm_gh.md).
 
+## Contributing
+
+Contributions in the form of feedback, comments, code, and bug report are welcome.
+
+* Fork the source code, modify, and issue a [pull request](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request-from-a-fork).
+* Issues, bug reports: [File a Github issue](https://github.com/chainsawriot/oolong).
+* Github is not your thing? Contact Chung-hong Chan by e-mail, post, or other methods listed on this [page](https://www.mzes.uni-mannheim.de/d7/en/profiles/chung-hong-chan).
+
+## Code of Conduct
+
+Please note that the oolong project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+
 ---
