### 2020-07-21 1fe8d8e
Update paper to have a summary of the field, ref #31
