+++ paper/paper.bib
@@ -139,3 +139,141 @@
     pages = {1--40},
     doi = {10.18637/jss.v091.i02},
   }
+
+@article{dimaggio2013exploiting,
+  title={Exploiting affinities between topic modeling and the sociological perspective on culture: Application to newspaper coverage of US government arts funding},
+  author={DiMaggio, Paul and Nag, Manish and Blei, David},
+  doi={10.1016/j.poetic.2013.08.004},
+  journal={Poetics},
+  volume={41},
+  number={6},
+  pages={570--606},
+  year={2013},
+  publisher={Elsevier}
+}
+
+@article{grimmer2011general,
+  title={General purpose computer-assisted clustering and conceptualization},
+  author={Grimmer, Justin and King, Gary},
+  journal={Proceedings of the National Academy of Sciences},
+  doi={10.1073/pnas.1018067108},
+  volume={108},
+  number={7},
+  pages={2643--2650},
+  year={2011},
+  publisher={National Acad Sciences}
+}
+
+@article{quinn2010analyze,
+  title={How to analyze political attention with minimal assumptions and costs},
+  author={Quinn, Kevin M and Monroe, Burt L and Colaresi, Michael and Crespin, Michael H and Radev, Dragomir R},
+  journal={American Journal of Political Science},
+  doi={10.1111/j.1540-5907.2009.00427.x},
+  volume={54},
+  number={1},
+  pages={209--228},
+  year={2010},
+  publisher={Wiley Online Library}
+}
+
+@Manual{jones2019textminer,
+  title = {textmineR: Functions for Text Mining and Topic Modeling},
+  author = {Tommy Jones},
+  year = {2019},
+  note = {R package version 3.0.4},
+  url = {https://CRAN.R-project.org/package=textmineR},
+}
+
+@Manual{schwemmer2018stminsights,
+  title = {stminsights: A 'Shiny' Application for Inspecting Structural Topic Models},
+  author = {Carsten Schwemmer},
+  year = {2018},
+  note = {R package version 0.3.0},
+  url = {https://CRAN.R-project.org/package=stminsights},
+}
+
+@Manual{sievert2015ldavis,
+  title = {LDAvis: Interactive Visualization of Topic Models},
+  author = {Carson Sievert and Kenny Shirley},
+  year = {2015},
+  note = {R package version 0.3.2},
+  url = {https://CRAN.R-project.org/package=LDAvis},
+}
+
+@Manual{koppers2020tosca,
+  title = {tosca: Tools for Statistical Content Analysis},
+  author = {Lars Koppers and Jonas Rieger and Karin Boczek and Gerret {von Nordheim}},
+  year = {2020},
+  url = {https://github.com/Docma-TU/tosca},
+  doi = {10.5281/zenodo.3591068},
+  note = {R package version 0.2-0},
+}
+
+@inproceedings{mimno2011optimizing,
+  title={Optimizing semantic coherence in topic models},
+  author={Mimno, David and Wallach, Hanna and Talley, Edmund and Leenders, Miriam and McCallum, Andrew},
+  booktitle={Proceedings of the 2011 Conference on Empirical Methods in Natural Language Processing},
+  pages={262--272},
+  year={2011}
+}
+
+@article{maier2018applying,
+  title={Applying LDA topic modeling in communication research: Toward a valid and reliable methodology},
+  author={Maier, Daniel and Waldherr, Annie and Miltner, Peter and Wiedemann, Gregor and Niekler, Andreas and Keinert, Alexa and Pfetsch, Barbara and Heyer, Gerhard and Reber, Ueli and H{\"a}ussler, Thomas and others},
+  doi={10.1080/19312458.2018.1430754},
+  journal={Communication Methods and Measures},
[…]
