+++ paper/paper.md
@@ -31,9 +31,21 @@ Oolong is an R package providing functions for validating common automated conte
 
 # Statement of need
 
-Validity is a requirement of content analysis [@krippendorff2018content; @neuendorf2016content]. Validation of automated methods has been called for by many scholars, e.g. @grimmer2013text; @ribeiro2016sentibench; @van2018communication. However, the method for validating these automated methods is unstandardized.
+Validity is a requirement of content analysis [@krippendorff2018content; @neuendorf2016content]. Validation of automated methods has been called for by many scholars, e.g. @grimmer2013text; @ribeiro2016sentibench; @van2018communication. 
 
 Oolong makes it easy to generate standard validation tests suggested by @chang2009reading and @song2020validations.
+
+# Validation of automated content analysis
+
+The paper by @dimaggio2013exploiting conceptualizes validation of automated methods as three different operations and the three operations supplement each other. These three operations are: 1) *statistical* validation -- to see if the model results agree with the assumptions of the model. Examples of statistical validation are calculation of pointwise mutual information, perplexity or semantic coherence of a topic model. 2) *semantic* validation -- to see if the model results are semantically meaningful. This procedure involves comparing model results with human judgment [@grimmer2011general]. 3) *predictive* validation -- to see if the model results can predict external events [@quinn2010analyze]. For example, one can study whether external events can explain surges in attention to a topic extracted by a topic model.
+
+This package focuses on semantic validation. The reason is threefold. First, there are existing architecture for conducting statistical validation and predictive validation. Topic modeling packages such as `text2vec` [@selivanov2020tex2vec] and `topicmodels` [@bettina2011topicmodels] provide functions to calculate metrics such as perplexity. Packages such as `textmineR` [@jones2019textminer], `stminsights` [@schwemmer2018stminsights] and `LDAvis` [@sievert2015ldavis] offers additional methods for statistical validation and predictive validation. As of writing, `tosca` [@koppers2020tosca] is the only package dealing with semantic validation. But the text-based interface might pose challenges to human annotators and it can only support topic models from one package [@change2015lda].
+
+Second, results from statistical validation do not always agree with those from semantic validation. For example, topics from a topic model with a lower perplexity do not have a better interpretability [@chang2009reading]. Of course, there are also metrics from statistical validation that are shown to be correlated with semantic validity, e.g. semantic coherence [@mimno2011optimizing]. Calculation of semantic coherence is recommended in the best practice paper by @maier2018applying. Nonetheless, conducting only statistical validation is not adequate because 3 validation operations supplement each other.
+
+Finally, predictive validation is dependent on research questions and thus it is difficult to be generalized as a reusable software framework. Additionally, the relationship between external (sociopolitical) events and the results from automated content analysis tools is usually what social scientists are eager to study, c.f. using topic models for information retrieval [@yi2008evaluating]. We do not believe social scientists would forget to conduct any form of predictive validation for their topic models.
+
[…]
