### 2020-07-21 f027ea5
fix small grammatical problems, ref #31
