### 2020-07-21 d64a383
port Marius's idea back to paper.rmd, ref #31
