+++ paper/paper.bib
@@ -277,3 +277,51 @@
     note = {R package version 1.4.2},
     url = {https://CRAN.R-project.org/package=lda},
   }
+
+@article{gonzalez2015signals,
+  title={Signals of public opinion in online communication: A comparison of methods and data sources},
+  author={Gonz{\'a}lez-Bail{\'o}n, Sandra and Paltoglou, Georgios},
+  journal={The ANNALS of the American Academy of Political and Social Science},
+  doi={10.1177/0002716215569192},
+  volume={659},
+  number={1},
+  pages={95--107},
+  year={2015},
+  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+}
+
+@article{boukes2020whatsthetone,
+  title={What’s the tone? Easy doesn’t do it: Analyzing performance and agreement between off-the-shelf sentiment analysis tools},
+  author={Boukes, Mark and van de Velde, Bob and Araujo, Theo and Vliegenthart, Rens},
+  journal={Communication Methods and Measures},
+  doi={10.1080/19312458.2019.1671966},
+  volume={14},
+  number={2},
+  pages={83--104},
+  year={2020},
+  publisher={Taylor \& Francis}
+}
+
+@article{boumans2016taking,
+  title={Taking stock of the toolkit: An overview of relevant automated content analysis approaches and techniques for digital journalism scholars},
+  author={Boumans, Jelle W and Trilling, Damian},
+  journal={Digital journalism},
+  doi={10.1080/21670811.2015.1096598},
+  volume={4},
+  number={1},
+  pages={8--23},
+  year={2016},
+  publisher={Taylor \& Francis}
+}
+
+@article{gunther2016word,
+  title={Word counts and topic models: Automated text analysis methods for digital journalism research},
+  author={G{\"u}nther, Elisabeth and Quandt, Thorsten},
+  journal={Digital Journalism},
+  doi={10.1080/21670811.2015.1093270},
+  volume={4},
+  number={1},
+  pages={75--88},
+  year={2016},
+  publisher={Taylor \& Francis}
+}
