### 2020-07-22 951b117
adding rhetorical question, ref #31
