### 2020-07-22 #2461 @chainsawriot
Dear Dr Waggoner @pdwaggoner ,

Thank you very much for your valuable comments.

My coauthor @msaeltzer and I have revised both the paper and the GitHub repo to address your comments. It can be reflected in the current version of chainsawriot/oolong@a0fa6d0.

* We have updated the [Github repo of oolong](https://github.com/chainsawriot/oolong) to provide a clear path to community contribution. We have provided a short guide to contribution and contributer code of conduct.
* We have addressed the two other points by you under the [section "Validation of Automated Content Analysis" of the paper](https://github.com/chainsawriot/oolong/blob/master/paper/paper.md#validation-of-automated-content-analysis). In this section, we outline what does it mean by validation of automated methods (Statistical validation, Semantic Validation, and Predictive Validation). The reason for us to focus on semantic validation is also provided.

We hope we have addressed all your concerns with both the Github repo and the paper. We are looking forward for your further feedback.

Regards,
Dr Chung-hong Chan @chainsawriot
