### 2020-07-22 #2461 @pdwaggoner
Hi @chainsawriot - Thanks for the revisions! The paper looks much better now, with a very clear picture of how your substantive and technical approaches fit in the broader literature. I also like how you've included code of conduct in addition to contribution guidelines. Great changes and solid software! This is good to go in my opinion.

Back to you @kakiac .
