+++ paper/paper.bib
@@ -325,3 +325,15 @@
   year={2016},
   publisher={Taylor \& Francis}
 }
+
+@article{fan2019assessing,
+  title={Assessing topic model relevance: Evaluation and informative priors},
+  author={Fan, Angela and Doshi-Velez, Finale and Miratrix, Luke},
+  journal={Statistical Analysis and Data Mining: The ASA Data Science Journal},
+  volume={12},
+  number={3},
+  pages={210--222},
+  year={2019},
+  publisher={Wiley Online Library},
+  doi={10.1002/sam.11415}
+}
