### 2020-08-02 a0fe10d
Edits related to the publication of Chan & Grill
