+++ paper/paper.bib
@@ -58,7 +58,7 @@
   author={Chan, Chung-hong and Grill, Christiane},
   journal={Communication Research},
   year={2020},
-  volume={forthcoming}
+  doi={10.1177/0093650220944790}
 }
 
 @article{song2020validations,
