+++ man/abstracts.Rd
@@ -23,4 +23,7 @@ abstracts_dictionary
 \description{
 This is a random sample of all abstracts of papers published in high-impact communication journals from 2000 to 2017. The one ends with "dfm" is the same data in quanteda::dfm (document-feature matrix) format. abstracts_dictionary is a list of terms that can be used for semisupervised techniques such as keyATM.
 }
+\references{
+Chan, C-h, & Grill, C. (2020). [The Highs in Communication Research: Research Topics With High Supply, High Popularity, and High Prestige in High-Impact Journals.](https://doi.org/10.1177/0093650220944790) Communication Research.
+}
 \keyword{datasets}
