### 2020-08-17 #2461 @chainsawriot
Dear Professor O'Donnell (@mbod) & @kakiac ,

I have made a minor update to the manuscript to reflect the changes in the bibliography (A cited paper is no longer "in press"). I hope it can facilitate the review. Thank you very much!

Regards,
Chung-hong Chan
