### 2020-08-30 #2461 @chainsawriot
Dear @kakiac ,

According to the blog post about the current [reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss), a review is expected to be finished in 6 weeks (was [2 weeks](https://joss.readthedocs.io/en/latest/editing.html)). If it takes longer than that, you, as an editor, is expected to intervene. According to my calendar (probably yours too), it has been 7 weeks. I was wondering if you can either check with the reviewer (@mbod ) or assign a new reviewer if @mbod is not available. Thank you very much!

cc. my coauthor @msaeltzer & the original EiC @danielskatz

Regards,
Chung-hong Chan
