### 2020-10-06 #2461 @Kudusch
Thank you as well to @kakiac for the chance to review for JOSS!

Dear @chainsawriot,

first of all: Thank you for writing this package! It has been very useful for me and I'm sure it'll be/it has been very useful for the community as well.

I have now read the paper and (re-)installed oolong and went through some of the examples provided by you. I think the paper and the package is almost fit for publication, I just have a few little comments:

- There is an issue with the R output in the paper on pages 3, 4, & 5. The lines seem to be to long and they are not being broken correctly. This is mainly an aesthetic issue, as the code itself is readable. Maybe, the output can be truncated for the paper?
- In the installation guidelines in the [overview](https://github.com/chainsawriot/oolong/blob/master/overview_gh.md) you recommend installing the package from GitHub. Is the developmental version the recommended version of the package or are installations through CRAN fine as well?
- In the paper on page 4 the output of the `summarize_oolong()` function might need a little more explanation; a guide on how these values can be interpreted. This information is (in part) also found in the package's documentation, but a paragraph or two on the values returned by `summarize_oolong()` might be helpful.

With these minor issues addressed, I think the paper can be published!
