### 2020-10-06 chainsawriot/oolong#31 @chainsawriot
Comment from @Kudusch

- [ ] There is an issue with the R output in the paper on pages 3, 4, & 5. The lines seem to be to long and they are not being broken correctly. This is mainly an aesthetic issue, as the code itself is readable. Maybe, the output can be truncated for the paper?
- [ ]In the installation guidelines in the overview you recommend installing the package from GitHub. Is the developmental version the recommended version of the package or are installations through CRAN fine as well?
- [ ] In the paper on page 4 the output of the summarize_oolong() function might need a little more explanation; a guide on how these values can be interpreted. This information is (in part) also found in the package's documentation, but a paragraph or two on the values returned by summarize_oolong() might be helpful.
