### 2020-10-13 c4588ed
Fixing recommendation about which version to install, ref #31
