+++ overview_gh.md
@@ -38,11 +38,11 @@ Models from `BTM` and Keyword Assisted Topic Models from `keyATM`.
 ``` r
 library(oolong)
 library(stm)
-#> stm v1.3.5 successfully loaded. See ?stm for help. 
+#> stm v1.3.6 successfully loaded. See ?stm for help. 
 #>  Papers, resources, and other materials at structuraltopicmodel.com
 library(quanteda)
-#> Package version: 2.0.1
-#> Parallel computing: 2 of 4 threads used.
+#> Package version: 2.1.2
+#> Parallel computing: 2 of 8 threads used.
 #> See https://quanteda.io for tutorials and examples.
 #> 
 #> Attaching package: 'quanteda'
@@ -309,7 +309,7 @@ abstracts_dfm
 
 ``` r
 oolong_test <- create_oolong(abstracts_warplda, abstracts$text, input_dfm = abstracts_dfm)
-#> INFO  [15:31:53.173] early stopping at 50 iteration
+#> INFO  [16:57:17.437] early stopping at 50 iteration
 oolong_test
 #> An oolong test object with k = 20, 0 coded.
 #> Use the method $do_word_intrusion_test() to do word intrusion test.
