### 2020-10-13 3596e86
Fixing recommendation about which version to install, ref #31
