### 2020-10-13 8d60748
Update vignette to reflect acceptance of Chan et al. 2020
