+++ overview_gh.md
@@ -316,7 +316,7 @@ abstracts_dfm
 
 ``` r
 oolong_test <- create_oolong(abstracts_warplda, abstracts$text, input_dfm = abstracts_dfm)
-#> INFO  [17:10:40.130] early stopping at 50 iteration
+#> INFO  [17:37:58.107] early stopping at 50 iteration
 oolong_test
 #> An oolong test object with k = 20, 0 coded.
 #> Use the method $do_word_intrusion_test() to do word intrusion test.
@@ -505,7 +505,7 @@ correlation and agreement.
   - Subplot (bottom left): Raw correlation between target value and
     content length. One should want to have no correlation, as an
     indication of good reliability against the influence of content
-    length. (See Chan et al.)
+    length. (See Chan et al. 2020)
   - Subplot (bottom right): Cook’s distance of all data point. One
     should want to have no dot (or at least very few dots) above the
     threshold. It is an indication of how the raw correlation between
@@ -536,16 +536,20 @@ plot(res)
 1.  Chang, J., Gerrish, S., Wang, C., Boyd-Graber, J. L., & Blei, D. M.
     (2009). Reading tea leaves: How humans interpret topic models. In
     Advances in neural information processing systems (pp. 288-296).
+    [link](https://papers.nips.cc/paper/3700-reading-tea-leaves-how-humans-interpret-topic-models)
 2.  Song et al. (2020) In validations we trust? The impact of imperfect
     human annotations as a gold standard on the quality of validation of
     automated content analysis. Political Communication.
-    [link](http://www.hyunjinsong.com/supplements/PC_Song_et_al_2020.pdf)
+    [link](https://doi.org/10.1080/10584609.2020.1723752)
 3.  Bland, J. M., & Altman, D. (1986). Statistical methods for assessing
     agreement between two methods of clinical measurement. The lancet,
     327(8476), 307-310.
-4.  Chan et al. (2020) It is not listed here, because it is still under
-    review. Do it faster\! Reviewers\!
+4.  Chan et al. (2020) Four best practices for measuring news sentiment
+    using ‘off-the-shelf’ dictionaries: a large-scale p-hacking
+    experiment. Computational Communication Research.
+    [link](https://osf.io/preprints/socarxiv/np5wa/)
 5.  Nielsen, F. Å. (2011). A new ANEW: Evaluation of a word list for
     sentiment analysis in microblogs. arXiv preprint arXiv:1103.2903.
+    [link](https://arxiv.org/abs/1103.2903)
 
 -----

[…]
