### 2020-10-14 77a61e8
Making printing oolong summary smarter, ref #31
