### 2020-10-14 chainsawriot/oolong#31 @kakiac
Hello @chainsawriot just linking to the original JOSS review issue, so that we can see the updates there as well: https://github.com/openjournals/joss-reviews/issues/2461
