### 2020-10-15 0a78277
Update paper to explain the diagnostic plot, ref #31
