+++ paper/paper.bib
@@ -1,3 +1,13 @@
+@article{chan_4b,
+ title={Four best practices for measuring news sentiment using ‘off-the-shelf’ dictionaries: a large-scale p-hacking experiment},
+ url={osf.io/preprints/socarxiv/np5wa},
+ DOI={10.31235/osf.io/np5wa},
+ journal={Computational Communication Research},
+ author={Chan, Chung-hong and Bajjalieh, Joseph and Auvil, Loretta and Wessler, Hartmut and Althaus, Scott and Welbers, Kasper and van Atteveldt, Wouter and Jungblut, Marc},
+ year={2020},
+ note={(in press)}
+}
+
 @article{eshima2020keyatm,
     title = {Keyword Assisted Topic Models},
     author = {Shusei Eshima and Kosuke Imai and Tomoya Sasaki},
