+++ paper/paper.md
@@ -17,7 +17,7 @@ affiliations:
  - name: Mannheimer Zentrum für Europäische Sozialforschung, Universität Mannheim
    index: 1
 citation_author: Chan & Sältzer.
-date: 3 July 2020
+date: 15 October 2020
 year: 2020
 bibliography: paper.bib
 output: rticles::joss_article
@@ -117,7 +117,7 @@ oolong_test
 ## An oolong test object with k = 20, 20 coded.
 ## 95%  precision
 ## With 25 cases of topic intrusion test. 25 coded.
-## TLO: -0.162
+## TLO: -0.129
 ```
 
 The suggested workflow is to have at least two human raters to do the same set of tests. Test object can be cloned to allow multiple raters to do the test. More than one test object can be studied together using the function `summarize_oolong()`.
@@ -150,18 +150,18 @@ summarize_oolong(oolong_test_rater1, oolong_test_rater2)
 ```
 
 ```
-## Mean model precision: 0.25
-## Quantiles of model precision: 0.1, 0.175, 0.25, 0.325, 0.4
+## Mean model precision: 0.45
+## Quantiles of model precision: 0.25, 0.35, 0.45, 0.55, 0.65
 ## P-value of the model precision
-##  (H0: Model precision is not better than random guess): 0.0551
-## Krippendorff's alpha: 0.22
+##  (H0: Model precision is not better than random guess): 0
+## Krippendorff's alpha: 0.015
 ## K Precision:
-## 0, 0, 0, 0, 0, 1, 0, 0, 0.5, 1, 0.5, 0, 0, 0.5, 0, 0, 0.5, 0.5, 0, 0.5
-## Mean TLO: -2.46
-## Median TLO: -3.27
-## Quantiles of TLO: -6.19, -4, -3.27, 0, 0
+## 0, 0.5, 1, 0, 1, 0.5, 1, 0, 0.5, 1, 0.5, 0, 0, 0.5, 0.5, 0.5, 0.5, 0.5, 0, 0.5
+## Mean TLO: -2.18
+## Median TLO: -2.31
+## Quantiles of TLO: -4.85, -3.63, -2.31, -0.13, 0
 ## P-Value of the median TLO 
-## (H0: Median TLO is not better than random guess): 0.714
+## (H0: Median TLO is not better than random guess): 0.3047
 ```
 
 # Semantic validation of dictionary-based methods
@@ -271,7 +271,7 @@ res
 ## Effect of content length: -0.323 (p = 0.1643)
 ```
 
-Additional diagnostic plots can also be displayed.
+Additional diagnostic plots can also be displayed (Figure 2).
 
 
 ```r
@@ -280,7 +280,12 @@ plot(res)
 
 ![Diagnostic plots generated by oolong](paper_files/figure-latex/diagplot-1.pdf) 
 
-The 4 subplots from left to right, top to bottom are: 1) correlation between human judgement and target value; 2) Bland-Altman plot; 3) correlation between target value and content length and 4) Cook's distance of all data point. These plots are helpful to determine criterion validity, agreement, robustness against content length and outliers of the target value.
+The 4 subplots from left to right, top to bottom are: 
+
+1. Correlation between human judgement and target value - A strong correlation between the two is an indicator of criterion validity of the target value.
+2. Bland-Altman plot - If the dots are randomly scattering around the mean value (solid line), it is an indicator of good agreement between human judgement and the target value.
+3. Correlation between target value and content length - If there is no strong correlation between the target value and content length, it is an indicator of robustness against the influence of content length [see @chan_4b].
+4. Cook's distance of all data points - if there are only a few dots above the threshold (dotted line), it is an indicator of robustness against the influence of outliers.
 
 # Acknowledgements
 
