+++ paper/paper.rmd
@@ -17,7 +17,7 @@ affiliations:
  - name: Mannheimer Zentrum für Europäische Sozialforschung, Universität Mannheim
    index: 1
 citation_author: Chan & Sältzer.
-date: 3 July 2020
+date: 15 October 2020
 year: 2020
 bibliography: paper.bib
 output: rticles::joss_article
@@ -265,13 +265,18 @@ Printing the summary shows Krippendorff's Alpha, an indicator of interrater reli
 res
 ```
 
-Additional diagnostic plots can also be displayed.
+Additional diagnostic plots can also be displayed (Figure 2).
 
 ```{r diagplot, fig.cap = "Diagnostic plots generated by oolong"}
 plot(res)
 ```
 
-The 4 subplots from left to right, top to bottom are: 1) correlation between human judgement and target value; 2) Bland-Altman plot; 3) correlation between target value and content length and 4) Cook's distance of all data point. These plots are helpful to determine criterion validity, agreement, robustness against content length and outliers of the target value.
+The 4 subplots from left to right, top to bottom are: 
+
+1. Correlation between human judgement and target value - A strong correlation between the two is an indicator of criterion validity of the target value.
+2. Bland-Altman plot - If the dots are randomly scattering around the mean value (solid line), it is an indicator of good agreement between human judgement and the target value.
+3. Correlation between target value and content length - If there is no strong correlation between the target value and content length, it is an indicator of robustness against the influence of content length [see @chan_4b].
+4. Cook's distance of all data points - if there are only a few dots above the threshold (dotted line), it is an indicator of robustness against the influence of outliers.
 
 # Acknowledgements
 
