### 2020-10-15 #2461 @chainsawriot
Dear Mr. Tim Schatto-Eckrodt (@Kudusch ),

Thank you so much for accepting the review and your tremendously useful comments. Regarding your three points, I have made the following changes:

1. We have followed your advice to truncate some output, e.g. from tibble. However, we want to keep the output from oolong to show how it guides us interactively. The output problem was partly an issue of the verbose output from `summarize_oolong`. We have made the output from `summarize_oolong` cleaner (e.g. rounding some numbers).

2. We have edited the [overview](https://github.com/chainsawriot/oolong/blob/master/overview_gh.md) to reflect our recommendation: We suggest using the development version because the package is constantly changing. However, we have also included the instruction to install the so-called "stable" (a.k.a. older) version from CRAN.

3. In the software paper, we have followed your advice to provide guidance on how to interpret the results from `summarize_oolong`.

On top of these, we have also updated some references to reflect the up-to-date version of some cited papers.

We hope we have addressed all your concerns. We are looking forward for your further feedback. Once again, thank you very much!

Regards,
Dr Chung-hong Chan
