### 2020-10-15 #2461 @Kudusch
Dear Chung-hong Chan (@chainsawriot),

perfect! I just looked through the latest article proof and the changes you made to the repo. You have addressed all of my comments and I have nothing left to complain about!

Thank you for the work you put into this package and the review process!

@kakiac I think, after the two review phases, the paper is now fit for publication and I recommend to accept it.
