### 2020-10-15 #2461 @mbod
@chainsawriot

I was not able to get all the tests to pass and didn't immediately see documentation on the scope of the tests and how to run them.

```
  [1] "0020-06-04" "0020-06-04" "0020-05-28" "0020-05-28" "0020-05-28" "0020-05-28" "0020-05-23" "0020-05-23" "0020-05-19"
 [10] "0020-05-12" "0020-05-12" "0020-05-12" "0020-05-12" "0020-05-12" "0020-05-11" "0020-05-08" "0020-05-08" "0020-05-08"
 [19] "0020-05-07" "0020-05-07" "0020-05-06" "0020-05-06" "0020-05-06" "0020-05-06" "0020-05-05" "0020-05-05" "0020-05-05"
 [28] "0020-05-05" "0020-05-05" "0020-05-05" "0020-05-04" "0020-05-04" "0020-05-04" "0020-05-04" "0020-05-02" "0020-05-02"
 [37] "0020-05-02" "0020-04-30" "0020-04-29" "0020-04-29" "0020-04-28" "0020-04-28" "0020-04-23" "0020-04-22" "0020-04-22"
 [46] "0020-04-22" "0020-04-22" "0020-04-22" "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-21"
 [55] "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-21" "0020-04-20" "0020-04-20" "0020-04-20" "0020-04-20"
 [64] "0020-04-20" "0020-04-20" "0020-04-20" "0020-04-20" "0020-04-20" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
 [73] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
 [82] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
 [91] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
[100] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
[109] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
[118] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17"
[127] "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-17" "0020-04-15"
[136] "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15"
[145] "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-15" "0020-04-14" "0020-04-14"
[154] "0020-04-14" "0020-04-14" "0020-04-14" "0020-04-14" "0020-04-14" "0020-04-14"
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%Y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%Y')),
+                         percent=releases/nrow(repr.df)) %>%
+              kable() %>%
+              kable_styling(full_width = F)
>
>
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=releases/nrow(repr.df)) %>%
+              kable() %>%
+              kable_styling(full_width = F)
>
>
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               reorder(percent) %>%
+              kable() %>%
+              kable_styling(full_width = F)
Error in tapply(X = X, INDEX = x, FUN = FUN, ...) :
  object 'percent' not found
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               order_by(percent) %>%
+              kable() %>%
+              kable_styling(full_width = F)
Error: `call` must be a function call, not a symbol
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               arrange(percent) %>%
+              kable() %>%
+              kable_styling(full_width = F)
>
>
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               arrange(percent, descending=F) %>%
+              kable() %>%
+              kable_styling(full_width = F)
Error: incorrect size (1) at position 2, expecting : 22
> ?arrange
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               arrange(percent, desc()) %>%
+              kable() %>%
+              kable_styling(full_width = F)
Error: cannot arrange column of class 'NULL' at position 2
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(releases=n(),
+                         first=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         last=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         percent=round(releases/nrow(repr.df),2)) %>%
+               arrange(desc(percent)) %>%
+              kable() %>%
+              kable_styling(full_width = F)
>
>
> max(as.Date(as.character(repr.df$Reprieve.Date)))
[1] NA
> max(as.Date(as.character(repr.df$Reprieve.Date, '%m/%d/%y')))
[1] NA
> as.Date(as.character(repr.df$Reprieve.Date, '%m/%d/%y'))
  [1] "0006-04-20" "0006-04-20" NA           NA           NA           NA           NA           NA           NA
 [10] "0005-12-20" "0005-12-20" "0005-12-20" "0005-12-20" "0005-12-20" "0005-11-20" "0005-08-20" "0005-08-20" "0005-08-20"
 [19] "0005-07-20" "0005-07-20" "0005-06-20" "0005-06-20" "0005-06-20" "0005-06-20" "0005-05-20" "0005-05-20" "0005-05-20"
 [28] "0005-05-20" "0005-05-20" "0005-05-20" "0005-04-20" "0005-04-20" "0005-04-20" "0005-04-20" "0005-02-20" "0005-02-20"
 [37] "0005-02-20" NA           NA           NA           NA           NA           NA           NA           NA
 [46] NA           NA           NA           NA           NA           NA           NA           NA           NA
 [55] NA           NA           NA           NA           NA           NA           NA           NA           NA
 [64] NA           NA           NA           NA           NA           NA           NA           NA           NA
 [73] NA           NA           NA           NA           NA           NA           NA           NA           NA
 [82] NA           NA           NA           NA           NA           NA           NA           NA           NA
 [91] NA           NA           NA           NA           NA           NA           NA           NA           NA
[100] NA           NA           NA           NA           NA           NA           NA           NA           NA
[109] NA           NA           NA           NA           NA           NA           NA           NA           NA
[118] NA           NA           NA           NA           NA           NA           NA           NA           NA
[127] NA           NA           NA           NA           NA           NA           NA           NA           NA
[136] NA           NA           NA           NA           NA           NA           NA           NA           NA
[145] NA           NA           NA           NA           NA           NA           NA           NA           NA
[154] NA           NA           NA           NA           NA           NA
> as.Date(as.character(repr.df$Reprieve.Date), '%m/%d/%y')
  [1] "2020-06-04" "2020-06-04" "2020-05-28" "2020-05-28" "2020-05-28" "2020-05-28" "2020-05-23" "2020-05-23" "2020-05-19"
 [10] "2020-05-12" "2020-05-12" "2020-05-12" "2020-05-12" "2020-05-12" "2020-05-11" "2020-05-08" "2020-05-08" "2020-05-08"
 [19] "2020-05-07" "2020-05-07" "2020-05-06" "2020-05-06" "2020-05-06" "2020-05-06" "2020-05-05" "2020-05-05" "2020-05-05"
 [28] "2020-05-05" "2020-05-05" "2020-05-05" "2020-05-04" "2020-05-04" "2020-05-04" "2020-05-04" "2020-05-02" "2020-05-02"
 [37] "2020-05-02" "2020-04-30" "2020-04-29" "2020-04-29" "2020-04-28" "2020-04-28" "2020-04-23" "2020-04-22" "2020-04-22"
 [46] "2020-04-22" "2020-04-22" "2020-04-22" "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-21"
 [55] "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-21" "2020-04-20" "2020-04-20" "2020-04-20" "2020-04-20"
 [64] "2020-04-20" "2020-04-20" "2020-04-20" "2020-04-20" "2020-04-20" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
 [73] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
 [82] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
 [91] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
[100] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
[109] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
[118] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17"
[127] "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-17" "2020-04-15"
[136] "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15"
[145] "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-15" "2020-04-14" "2020-04-14"
[154] "2020-04-14" "2020-04-14" "2020-04-14" "2020-04-14" "2020-04-14" "2020-04-14"
> max(as.Date(as.character(repr.df$Reprieve.Date), '%m/%d/%y'))
[1] "2020-06-04"
> min(as.Date(as.character(repr.df$Reprieve.Date), '%m/%d/%y'))
[1] "2020-04-14"
Error in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]) :
  namespace ‘rmarkdown’ 1.15 is already loaded, but >= 1.17 is required
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
>
> # Here we show an initial descriptive analysis of the composition of these 156 inmates, in which we describe the racial makeup of this group. we find that the racial composition of this group roughly reaches parity with regard to the racial composition of prisoners in PA prisons (demographic population data extrapolated from 2018 data). See the table below for the breakdown of racial composition for reprieve and prison population.
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(Num.Releases=n(),
+                         First.Release=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         Last.Release=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         Percentage.of.Releases=round(releases/nrow(repr.df),2)) %>%
+               arrange(desc(percent)) %>%
+              kable() %>%
+              kable_styling(full_width = F)
Error: object 'releases' not found
> repr.df <- read.csv('../data/Reprive_data.csv')
>
>
>
> # Here we show an initial descriptive analysis of the composition of these 156 inmates, in which we describe the racial makeup of this group. we find that the racial composition of this group roughly reaches parity with regard to the racial composition of prisoners in PA prisons (demographic population data extrapolated from 2018 data). See the table below for the breakdown of racial composition for reprieve and prison population.
>
>
> # I think we should hold off on this analysis for now
> # it feels a little bit like using the race category just because it is in the data
> # needs a bit more thought
> #
> # c = repr.df %>% group_by(Race) %>%
> #   summarise(reprieve = round((n()/nrow(repr.df))*100)) %>%
> #   filter(Race %in% c("ASIAN",'BLACK',"WHITE",'HISPANIC'))
> #
> # c$population = c(1,47,9,43)
> # c %>%
> #   mutate(reprieve = paste(reprieve,"%",sep=''),
> #          population = paste(population,"%",sep='')) %>%
> #   kable() %>%
> #   kable_styling(full_width = F)
>
> # Think what we do want to highlight is the timespan of releases
> # and maybe the SCI
>
> repr.df %>%  group_by(SCI) %>%
+               summarise(Num.Releases=n(),
+                         First.Release=min(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         Last.Release=max(as.Date(as.character(Reprieve.Date), format='%m/%d/%y')),
+                         Percentage.of.Releases=round(Num.Releases/nrow(repr.df),2)) %>%
+               arrange(desc(Percentage.of.Releases)) %>%
+              kable() %>%
+              kable_styling(full_width = F)
>
>
Error in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]) :
  namespace ‘rmarkdown’ 1.15 is already loaded, but >= 1.17 is required
> unique([1,1,1])
Error: unexpected '[' in "unique(["
> unique(c(1,1,1))
[1] 1
> unique(c(1,1,1,2,2))
[1] 1 2
[…]
