### 2020-10-15 #2461 @mbod
@chainsawriot paper now reads well - installation steps are documentation.

* However, I would anticipate that a segment of the user-base who would really want to benefit from a tool like this might need more hand holding and walk through setup. Perhaps assuming all will be RStudio and R familiar is appropriate given the specific tools/models oolong is designed to interoperate with. For the future some more basic tutorials or even videos could be a strong addition to your very useful tool!

Thanks! Good work!
