### 2020-10-15 #2461 @chainsawriot
@mbod I think the de facto standard way to run unit tests of an R package is for you to clone the git repo to your local; change the working directory to the base directory of the R package; and then run: `devtools::test()` or if you like: `devtools::check(manual = TRUE, remote = TRUE)`. But please make sure you have set up an environment that can run those checks (e.g. latex etc) if you want to do `check`. It would be good also if you can provide your `sessionInfo()`

More information on how to run tests of an R package can be found in the book [R Packages](https://r-pkgs.org/tests.html#test-workflow). Also, some tests are skipped on CRAN, because they are testing the features depending on the packages listed in the `Suggested` field of the [DESCRIPTION](https://github.com/chainsawriot/oolong/blob/master/DESCRIPTION), e.g. `topicmodels`, `keyATM`, `BTM` etc. If you want to test it on your local machine, it is generally asummed to have all the suggested packages installed (the same is true for doing standard development tasks such as building vignettes). I think it has been a pretty standard convention already. I don't need to manually instruct [travis CI](https://github.com/chainsawriot/oolong/blob/master/.travis.yml), for example, to install all suggested packages to run all the tests. I believe the arrangement for GitHub Actions is pretty similar.

Could you please give it a try once again? Thank you very much.
