### 2020-10-15 #2461 @chainsawriot
@mbod Could you please also elaborate on what did you mean by "need more hand-holding and walk through setup", so that we can improve the paper accordingly?

Did you mean the information such as installation instructions or even information on how to train a topic model? This kind of hands on information can be found in the [overview on Github](https://github.com/chainsawriot/oolong/blob/master/overview_gh.md) or the [vignette](https://cran.r-project.org/web/packages/oolong/vignettes/overview.html).

This is my first submission to JOSS, maybe it is also a good opportunity for me to consult @kakiac on what kind of information should be include in a JOSS software paper. The current oolong paper is 4 times longer than the [example JOSS paper](https://www.theoj.org/joss-papers/joss.00388/10.21105.joss.00388.pdf). A more general text analysis R package facing probably the same audience of oolong is [quanteda](https://github.com/quanteda/quanteda). The [JOSS software paper](https://joss.theoj.org/papers/10.21105/joss.00774) of quanteda does not contain a lot of information on usage. A recently accepted paper of [textnet](https://joss.theoj.org/papers/10.21105/joss.02594) is the standard JOSS 3-pager.

The current software paper has overlapped a lot with the vignette already. Would a hyperlink from the software paper to the vignette be enough as a means to "hand-holding and walk through setup"?

@mbod 's advice on providing tutorials +/- videos has been implemented. For example, in the previous ICA conference, I have provided a tutorial at a preconference. A video has [been recorded](https://www.youtube.com/watch?v=ZFSR0aUaJS8). The video is too rough to be included in the README, however. In the future, we will seek opportunties to give workshops on validating text analysis.
