### 2020-10-19 #2461 @kakiac
Thank you all @chainsawriot @mbod @Kudusch @pdwaggoner - great work and the paper and repo look much improved as a result! @chainsawriot many thanks for taking on board the advice and recording the tutorial 👍

@mbod can you confirm that you are happy that the following two items in your checklist have been addressed:

[] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

@mbod, could you also confirm that you are happy to recommend acceptance of the paper?
