### 2020-10-19 #2461 @chainsawriot
@kakiac I would like to say that Whedon's suggested DOI for the Krippendorf's book is incorrect. The DOI suggested is [a book review of the book](https://www.jstor.org/stable/2288384?seq=1#metadata_info_tab_contents). As far as I know, the book has no DOI.
