### 2020-10-27 #2461 @chainsawriot
Dear @kakiac ,

I was wondering if the paper is now "conditionally accepted" or not?

Are we waiting for @mbod 's response? If yes, should we set a time limit for this?

Thank you very much!
