### 2020-11-05 #2461 @kakiac
> Dear @kakiac ,
>
> I was wondering if the paper is now "conditionally accepted" or not?
>
> Are we waiting for @mbod 's response? If yes, should we set a time limit for this?
>
> Thank you very much!

Thanks @chainsawriot! On the basis of at least two reviewers having ticked their checklists and the quality of the submission, I am happy to accept the paper, and proceed with the next steps. Can you please:
 - make a tagged release (on github) and
 - archive (on Zenodo or Figshare), and
 - report the tagged version number and archive DOI in this review thread.

Let me know if you need any help with either.

(**Note**: Upon successful completion of the review, you need to make a tagged release of the software (via github, see here for info https://git-scm.com/book/en/v2/Git-Basics-Tagging), and deposit a copy of the repository with a data-archiving service such as Zenodo or figshare.

You then need to get the DOI you will be assigned from them (it is automatically assigned once you create a repository).

As soon as you have both copy the release version number (from github) and the DOI in this thread.

See here for more info under section The review process https://joss.readthedocs.io/en/latest/submitting.html#submitting-your-paper - let me know if you need any help!

An example of a recent tagged release on github is here: https://github.com/kgoldfeld/simstudy/ archived on Zenodo with the DOI: https://doi.org/10.5281/zenodo.4134675)
