### 2020-11-07 #2461 @chainsawriot
Dear @kakiac

Thanks for the info. I have created a tagged release on both Github and Zenodo. It's version 0.3.11.

https://github.com/chainsawriot/oolong/releases/tag/0.3.11

https://doi.org/10.5281/zenodo.4256574

Please let me know what I should do next. Thank you very much!

Regards,
Chung-hong Chan
