### 2020-11-09 #2461 @kthyng
@chainsawriot Please update the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper.
