### 2020-11-09 #2461 @kthyng
@chainsawriot Just finished going through your paper. It looks good! Just one issue is that some items in your references that should be capitalized are not. To preserve capitalization in your .bib file, you can put {} around the word or a whole title if you prefer. Please look through your references in detail and make sure all appropriate items are capitalized.
