### 2020-11-09 #2461 @kakiac
> @chainsawriot Did you address the missing DOI above? I just did a quick check and it looks like it does indeed link to a publication of that title.

@kthyng: just to confirm, that the missing DOI offered by Whedon is actually referencing a review of the paper (rather than the paper itself) and there is no DOI we could find of the paper, unfortunately.
