### 2020-11-10 7be3eb4
Correcting bib; update README also
