+++ paper/paper.bib
@@ -3,7 +3,7 @@
  url={osf.io/preprints/socarxiv/np5wa},
  DOI={10.31235/osf.io/np5wa},
  journal={Computational Communication Research},
- author={Chan, Chung-hong and Bajjalieh, Joseph and Auvil, Loretta and Wessler, Hartmut and Althaus, Scott and Welbers, Kasper and van Atteveldt, Wouter and Jungblut, Marc},
+ author={Chan, Chung-hong and Bajjalieh, Joseph and Auvil, Loretta and Wessler, Hartmut and Althaus, Scott and Welbers, Kasper and Van Atteveldt, Wouter and Jungblut, Marc},
  year={2020},
  note={(in press)}
 }
@@ -36,7 +36,7 @@
 
 
 @manual{selivanov2020tex2vec,
-    title = {text2vec: Modern Text Mining Framework for {R}},
+    title = {{t}ext2vec: Modern Text Mining Framework for {R}},
     author = {Dmitriy Selivanov and Manuel Bickel and Qing Wang},
     year = {2020},
     note = {R package version 0.6},
@@ -44,7 +44,7 @@
 }
   
 @article{benoit2018quanteda,
-  title={quanteda: An {R} package for the quantitative analysis of textual data},
+  title={{quanteda}: An {R} package for the quantitative analysis of textual data},
   author={Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and M{\"u}ller, Stefan and Matsuo, Akitaka},
   journal={Journal of Open Source Software},
   volume={3},
@@ -56,7 +56,7 @@
 
 
 @article{nielsen2011new,
-  title={A new ANEW: Evaluation of a word list for sentiment analysis in microblogs},
+  title={A new {ANEW}: Evaluation of a word list for sentiment analysis in microblogs},
   author={Nielsen, Finn {\AA}rup},
   journal={arXiv preprint arXiv:1103.2903},
   year={2011}
@@ -92,7 +92,7 @@
 @inproceedings{chang2009reading,
   title={Reading tea leaves: How humans interpret topic models},
   author={Chang, Jonathan and Gerrish, Sean and Wang, Chong and Boyd-Graber, Jordan L and Blei, David M},
-  booktitle={Advances in neural information processing systems},
+  booktitle={{Advances in Neural Information Processing Systems}},
   pages={288--296},
   year={2009},
   url={https://papers.nips.cc/paper/3700-reading-tea-leaves-how-humans-interpret-topic-models}
@@ -151,7 +151,7 @@
   }
 
 @article{dimaggio2013exploiting,
-  title={Exploiting affinities between topic modeling and the sociological perspective on culture: Application to newspaper coverage of US government arts funding},
+  title={Exploiting affinities between topic modeling and the sociological perspective on culture: Application to newspaper coverage of {US} government arts funding},
   author={DiMaggio, Paul and Nag, Manish and Blei, David},
   doi={10.1016/j.poetic.2013.08.004},
   journal={Poetics},
@@ -195,7 +195,7 @@
 }
 
 @Manual{schwemmer2018stminsights,
-  title = {stminsights: A 'Shiny' Application for Inspecting Structural Topic Models},
+  title = {{stminsights}: A '{S}hiny' Application for Inspecting Structural Topic Models},
   author = {Carsten Schwemmer},
   year = {2018},
   note = {R package version 0.3.0},
@@ -203,7 +203,7 @@
 }
 
 @Manual{sievert2015ldavis,
-  title = {LDAvis: Interactive Visualization of Topic Models},
+  title = {{LDAvis}: Interactive Visualization of Topic Models},
   author = {Carson Sievert and Kenny Shirley},
   year = {2015},
   note = {R package version 0.3.2},
[…]
