### 2020-11-10 #2461 @chainsawriot
Dear @kthyng and @kakiac

Thank you so much for going through the paper. Regarding the issues:

1. I have updated the [metadata in Zenodo](https://doi.org/10.5281/zenodo.4256574) so that the metadata matches both the paper, Github and CRAN.
2. As mentioned above, the missing DOI was a "false positive" by whedon. The alleged missing DOI links to the book review of the Krippendorff (a popular textbook on quantative content analysis). As far as I know, the Krippendorff doesn't have a DOI.
3. I have carefully updated the BibTeX file, especially on the capitalization, e.g. Abbreviations and proper nouns such as `US`, `LDA`, `ANEWS`,  `Gibbs sampling` are capitalized, whereas R package names, e.g. `lda`, `stminsights` etc are lowercased.

I hope with these revisions, the paper can be published.

Thank you very much!

Regards,
Chung-hong Chan
