### 2020-11-12 #2461 @arfon
@chainsawriot - No problem. I've just updated the paper and Crossref deposit metadata with your ORCID. This should be updated on the live site in the next hour or two.
